구버전(5.0.2.2 정도)에서 저장된 picture 컨트롤 읽을 때 오류 현상 수정.

+++ src/kr/dogfoot/hwplib/reader/ForUnknown.java
@@ -22,7 +22,7 @@ public class ForUnknown {
 	 */
 	public static void read(UnknownRecord unknown, StreamReader sr)
 			throws IOException {
-		byte[] body = new byte[unknown.getHeader().getSize()];
+		byte[] body = new byte[(int)unknown.getHeader().getSize()];
 		sr.readBytes(body);
 		unknown.setBody(body);
 	}
